Fix warning produced by building libotr
The warning: "A function declaration without a prototype is deprecated in all versions of C"

This should likely be better applied as a patch when building the library. I don't think that will be much of a concern seeing as the library was last updated in 2016. At worst, if these headers are replaced, it will just produce a warning.